Add an "explode" transport
Add a transport that explodes layers onto a directory tree somewhere on
the local filesystem.  Because that effectively destroys information
about multiple layers, we treat the result as an image with a single
layer as its contents.  This changes the manifest, which breaks any
signatures which have been computed on it.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>