Add instanceDigest to PutManifest/PutSignatures/LayerInfosForCopy
Add an instanceDigest parameter to PutManifest(), PutSignatures(), and
LayerInfosForCopy, for symmetry with GetManifest() and GetSignatures().
Return an error if the instanceDigest is supplied to destinations which
don't support them, and add stubs that do so even to the transports
which would support it, so that we don't break compilation here.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>