Improve detection of manifest list changes durin copy
When copying, switch from using an encode-then-compare of the original
and updated versions of the list to checking if the instance list was
changed (one of the things we might have changed) or if its type has
changed due to conversion (the other change we might have made).  If
neither is the same, then we don't need to change the encoded value of
the manifest.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>