image: drop local Schema2List type definitions
Drop local definitions of Schema2List types from the image package, in
favor of the ones that are now available from the manifest package.

Update manifestOCI1FromImageIndex(), manifestSchema2FromManifestList(),
and ChooseManifestInstanceFromManifestList() to use the ChooseDigest()
method that the manifest.ManifestList interface provides.

Revert to using image.ChooseManifestInstanceFromManifestList() in
copy.Image() to select a single image from a list, now that the former
knows about OCI types.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>