docker_image_dest: use blobinfocache2 compression info for reused blobs
Teach BlobInfoCache2 implementations to be able to retrieve the type of
compression that was applied to a blob with a given digest, if we know
it.

If we successfully reuse a blob that's already in the destination
repository while writing an image, and we know how a blob with that
reused blob's digest was compressed, use that information to update the
MIME type for the returned BlobInfo.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>