copy: use blobinfocache2 compression info for reused blobs
Teach BlobInfoCache2 implementations to be able to retrieve the type of
compression that was applied to a blob with a given digest, if we know
it.

If we successfully reuse a blob while writing an image, make sure we
update the returned BlobInfo with compression information so that the
MIME type in the updated image will be derived correctly.  When we're
copying a blob unmodified, return its BlobInfo with compression
information for the same reason.

In the containers-storage transport and while copying, when we
decompress a blob, record that its diffID is the digest of a blob that
we know is not compressed.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>